Fix invalid offset error in API response
It is possible for the API request to generate an invalid offset error if the array keys for status or body do not exist.

This fixes it by checking to ensure the keys exist before attempting to use them. It will generate a WP_Error if one or both of the keys are missing in the response.